Method/CPM/requirements.txt
Pillow==10.1.0
decord
sentencepiece==0.1.99
torch==2.1.2
torchvision==0.16.2
transformers==4.40.0
